Don't remove jurisdiction code when copying tax
Right now I think the most common use case for duplicating a tax is when
creating supplier purchase taxes, and it would be helpful to retain the
place and code. (This is subject to revision if my assumption turns out
    not to be the case.)

However, unset `place` and `code` when creating a generic tax.