Redesigning for one tax rule per state - naïve approach
Collapsed all tax rules of a given state into one tax rule using a
composite tax key to filter the correct taxes. This ended up with a lot
of rule lines that dealt with the same taxes.

Using Utah as an example, it meant moving 161 tax rules with 12 rule
lines each to one tax rule for the state with 1932 rule lines. Besides
the extra space required for storing the composite tax key with each
boundary (as a Char instead of the foreign key for the rule) it adds
about 0.36--0.5 seconds to the run the `TaxRuleLine.match` over that
many tax rule lines (first run from a sqlite on-disk test db; subsequent
calls run from the cache with negligible call time).

I thought of a better approach, but storing this approach here for
posterity. Yes, I'm looking at you. ;-) Because of this, I didn't bother
updating tests.