One rule - second approach (smarter)
Have one tax rule line per tax. Store the composite tax key on the
boundary and split it by a separator (hyphen) when searching for
matches. Locate via the code of the subsitution tax (in that case we
don't need to store the tax key on the tax rule line).

In order to save space when there will be potentially millions of
boundary records, I split the boundary table, saving the tax key in a
separate table since there are relatively few combinations of taxes
compared with boundary records, and some of the combinations lead to
long tax keys (12 bytes or more with separators).